[afin-dev] Refactored GitHub_RepoInspector to split it into GitHub_Client and a smaller GitHub_RepoInspector
