+++ src/conway_ops/repo_admin/github_repo_inspector.py
@@ -1,11 +1,9 @@
-import requests                                             as _requests
 from dateutil                                               import parser as _parser
 
 from conway.application.application                         import Application
 from conway.util.date_utils                                 import DateUtils
-from conway.util.yaml_utils                                 import YAML_Utils
 
-from conway_ops.repo_admin.github_response_handler          import GitHub_ReponseHandler
+from conway_ops.util.github_client                          import GitHub_Client
 from conway_ops.repo_admin.repo_inspector                   import RepoInspector, CommitInfo, CommittedFileInfo
 
 class GitHub_RepoInspector(RepoInspector):
@@ -31,133 +29,12 @@ class GitHub_RepoInspector(RepoInspector):
         # so we can extract the owner name from it ("alejandro-fin" in the example). This will be needed
         # to construct the URLs for other calls to the Git Hub API
         #
-        cleaned_url                                     = parent_url.strip("/").strip()
-        self.owner                                      = cleaned_url.split("/")[-1]
+        cleaned_url                             = parent_url.strip("/").strip()
+        github_owner                            = cleaned_url.split("/")[-1]
 
-        if len(self.owner) == 0:
-            raise ValueError("No owner was included in parent url '" + str(parent_url) + "', so can't call Git Hub APIs")
-        
-        # Get GitHub token to make authenticated calls.
-        #
-        SECRETS_PATH                                    = Application.app().config.secrets_path()
-        secrets_dict                                    = YAML_Utils().load(SECRETS_PATH)
-        self.github_token                               = secrets_dict['secrets']['github_token']
 
-        #self.github_api_url                             = f"https://{self.owner}@api.github.com"
-        self.github_api_url                             = f"https://api.github.com"
-    
-    def GET(self, resource, sub_path):
-        '''
-        Invokes the "GET" HTTP verb on the Git Hub API to get a resource associated to this inspector's repo.
-
-        :param str resource: indicates the top resource for the API. For example, "{owner}/repos". 
-        :param str sub_path: Indicates the path of a desired sub-resource to get, under the URL for the
-            `resource`. Examples: "/commits/master", "/branches", "/pulls" 
-
-        :return: A Json representation of the resource as given by the GitHub API
-        :rtype: str
-        '''
-        return self._http_call("GET", resource=resource, sub_path=sub_path, body={}, )
-    
-    def POST(self, resource, sub_path, body):
-        '''
-        Invokes the "POST" HTTP verb on the Git Hub API to create a resource associated to this inspector's repo.
-
-        :param str resource: indicates the top resource for the API. For example, "{owner}/repos". 
-        :param str sub_path: Indicates the path of a desired sub-resource to create, under the URL for the
-            `resource`. Examples: "/commits/master", "/branches", "/pulls" 
-        :param dict|list body: JSON object to submit in the HTTP request.
-        :return: A Json representation of the resource as given by the GitHub API
-        :rtype: str
-        '''
-        return self._http_call("POST", resource=resource, sub_path=sub_path, body=body)
-    
-    def PUT(self, resource, sub_path, body):
-        '''
-        Invokes the "PUT" HTTP verb on the Git Hub API to update a resource associated to this inspector's repo.
-
-        :param str resource: indicates the top resource for the API. For example, "{owner}/repos".
-        :param str sub_path: Indicates the path of a desired sub-resource to update, under the URL for the
-            `resource`. Examples: "/commits/master", "/branches", "/pulls" 
- 
-        :param dict|list body: JSON object to submit in the HTTP request.
-        :return: A Json representation of the resource as given by the GitHub API
-        :rtype: str
-        '''
-        return self._http_call("PUT", sub_path=sub_path, body=body, resource=resource)
-           
-    def DELETE(self, resource, sub_path):
-        '''
-        Invokes the "DELETE" HTTP verb on the Git Hub API to update a resource associated to this inspector's repo.
-
-        :param str resource: indicates the top resource for the API. For example, "{owner}/repos". 
-        :param str sub_path: Indicates the path of a desired sub-resource to delete, under the URL for the
-            `resource`. Examples: "/commits/master", "/branches", "/pulls" 
-
-        :param dict body: payload to submit in the HTTP request.
-        :return: A Json representation of the resource as given by the GitHub API
-        :rtype: str
-        '''
-        return self._http_call("DELETE", sub_path=sub_path, resource=resource)
-        
-    def _http_call(self, method, resource, sub_path, body={}):
-        '''
-        Invokes the Git Hub API to get information about the repo associated to this inspector.
-
-        :param str method: the HTTP verb to use ("GET", "POST", or "PUT")
-        :param str sub_path: Indicates the path of a desired sub-resource to delete, under the URL for the
-            `resource`. Examples: "/commits/master", "/branches", "/pulls" 
-        :param str resource: indicates the top resource for the API. For example, "repos". It is an optional
-            argument that defaults to "repos" if it is not provided.
-        :param dict body: optional payload to submit in the HTTP request. 
-        :return: A Json representation of the resource as given by the GitHub API
-        :rtype: str
-        '''
-        match resource:
-            case "repos":
-                root_path                   = f"{self.github_api_url}/repos/{self.owner}"
-            case "orgs":
-                root_path                   = f"{self.github_api_url}/orgs/{self.owner}"
-            case "users":
-                root_path                   = f"{self.github_api_url}/users/{self.owner}"
-            case "user":
-                # GOTCHA:
-                #       The "user" resource represents the currently authenticated user. As opposed to the "users"
-                #   resource, which can manipulate "other" users different from the currently authenticated user.
-                #   To create/update repos for a user, use the "user" resource, not the "users" resource.
-                root_path                   = f"{self.github_api_url}/user"
-            case "": # Return meta information
-                root_path                   = f"{self.github_api_url}"
-            case _:
-                raise ValueError(f"Unsupported GitHub resource '{resource}'")
-            
-        url                                 = root_path + sub_path
-
-        # Uncomment to debug
-        #Application.app().log(f"... calling '{method} {url}'")
-        
-        headers = {
-            'Authorization': 'Bearer ' + self.github_token,
-            'Content-Type' : 'application/json',
-            # GOTCHA:
-            #       Painfully found that GitHub post APIs will only work with the "vnd.github*" MIME types
-            #'Accept'       : 'application/json'
-            'Accept'        : 'application/vnd.github+json'
-            
-        }
-        try:
-            response                        = _requests.request(method          = method, 
-                                                                url             = url, 
-                                                                json            = body,
-                                                                headers         = headers, 
-                                                                timeout         = 20,
-                                                                verify          = True) 
-
-        except Exception as ex:
-            raise ValueError("Problem connecting to Git Hub. Error is: " + str(ex))
-        
-        return GitHub_ReponseHandler().process(response)    
-
+        self.github                             = GitHub_Client(github_owner = github_owner)
+     
     def current_branch(self):
         '''
         :return: The name of the current branch
@@ -205,7 +82,7 @@ class GitHub_RepoInspector(RepoInspector):
         :return: A :class:`CommitInfo` with information about last commit"
         :rtype: str
         '''
-        data                                = self.GET(resource = "repos",
+        data                                = self.github.GET(resource = "repos",
                                                        sub_path = f"/{self.repo_name}/commits/master")
         
         commit_datetime                     = _parser.parse(data['commit']['author']['date'])
@@ -229,7 +106,7 @@ class GitHub_RepoInspector(RepoInspector):
         :return: (local) branches for the repo
         :rtype: list[str]
         '''
-        data                                = self.GET(resource = "repos",
+        data                                = self.github.GET(resource = "repos",
                                                        sub_path = f"/{self.repo_name}/branches")
 
         result                              = [b['name'] for b in data]
@@ -242,7 +119,7 @@ class GitHub_RepoInspector(RepoInspector):
         (i.e., a log) for the repo associated to this :class:`RepoInspector`
         '''
         # This provides the first most recent commit, and links to "parent" commits - the commits right before it
-        data                                = self.GET(resource = "repos",
+        data                                = self.github.GET(resource = "repos",
                                                        sub_path = f"/{self.repo_name}/commits/master")
 
         results_dict                        = self._committed_files_impl(results_dict_so_far={}, data=data)
@@ -311,7 +188,7 @@ class GitHub_RepoInspector(RepoInspector):
                                                "base":      to_branch}
 
 
-        pr_result                           =  self.POST(
+        pr_result                           =  self.github.POST(
                                                         resource    = "repos",
                                                         sub_path    = f"/{self.repo_name}/pulls", 
                                                         body        = pr_data)
@@ -346,7 +223,8 @@ class GitHub_RepoInspector(RepoInspector):
                                                 "sha":              sha,
                                                 "merge_method":     "merge"}
 
-        merge_result                    =  self.PUT(    resource    = "repos",
+        merge_result                    =  self.github.PUT(    
+                                                        resource    = "repos",
                                                         sub_path    = f"/{self.repo_name}/pulls/{pull_number}/merge", 
                                                         body        = merge_data)
 
@@ -418,7 +296,7 @@ class GitHub_RepoInspector(RepoInspector):
         # Now do recursion, for each parent
         parents                             = data['parents']
         for p in parents:
-            p_data                          = self.GET(resource = "repos",
+            p_data                          = self.github.GET(resource = "repos",
                                                        sub_path = f"/{self.repo_name}/{p['url']}")
             results_dict                    = self._committed_files_impl(results_dict, p_data)
